remove from_dict and replace by an improved constructor
