+++ node-runner-cli/config/BaseConfig.py
@@ -1,8 +1,20 @@
+import yaml
+
+
 class BaseConfig:
     def __init__(self, settings: dict):
         if settings is not None:
             for key, value in settings.items():
                 setattr(self, key, value)
+
+            class_variables = {key: value
+                               for key, value in self.__class__.__dict__.items()
+                               if not key.startswith('__') and not callable(value)}
+            for attr, value in class_variables.items():
+                if type(self.__getattribute__(attr)) not in (str, int, bool, dict) and self.__getattribute__(
+                        attr) is not None:
+                    if (settings.get(attr) is not None):
+                        self.__getattribute__(attr).__init__(settings[attr])
 
     def __repr__(self):
         return repr(vars(self))
@@ -25,6 +37,16 @@ class BaseConfig:
                     attr) is not None:
                 returning_dict[attr] = self.__getattribute__(attr).to_dict()
         return returning_dict
+
+    def to_yaml(self):
+        config_to_dump = self.to_dict()
+        return yaml.dump(config_to_dump, sort_keys=True, default_flow_style=False, explicit_start=True,
+                         allow_unicode=True)
+
+    def to_file(self, config_file):
+        config_to_dump = self.to_dict()
+        with open(config_file, 'w') as f:
+            yaml.dump(config_to_dump, f, sort_keys=True, default_flow_style=False)
 
 
 class SetupMode:
